DRG: Drakesbane & new combo chain support
Dragoon was simplified to have Fang and Claw and Wheeling Thrust be
normal combo actions, and a new unique action was added to replace what
used to be the other one in the chain. Drakesbane is an action that
cannot be assigned to a bar, as the game usually swaps it itself, but
because of how this plugin works we have to handle it like any other
action.